pyproject.toml
[project]
name = "qq-pet-manager"
version = "0.1.0"
description = "QQ宠物（怀旧服）OpenClaw Skill - 自动养护、状态监控、疾病诊断"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pycryptodome>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
